+++ crates/rayexec_execution/src/hybrid/stream.rs
@@ -7,7 +7,7 @@ use rayexec_io::http::HttpClient;
 
 use super::client::{HybridClient, PullStatus};
 use crate::database::DatabaseContext;
-use crate::execution::intermediate::StreamId;
+use crate::execution::intermediate::pipeline::StreamId;
 use crate::execution::operators::sink::{PartitionSink, SinkOperation};
 use crate::execution::operators::source::{PartitionSource, SourceOperation};
 use crate::explain::explainable::{ExplainConfig, ExplainEntry, Explainable};
